little config + Debug.Log change
I moved the default values inside the class. I think it's a little bit
better. In the modloader, when it loads a mod with the mod attribute, it
says "Successfully loaded MyMod(v 0.1) by Letum\n
Successfully loaded MyMod.dll", so I moved the second thing that should
be said when there is no attribute inside of the if statement.