Overhaul of the mod loader, Part 1
This integrates many improvements from Vapid's Mod Loader,
which will not be developed further.
Most of the API for Mods was revamped, including the way mods are loaded.
The interface of the console was improved massively, thanks to work done by
VapidLinus. The object explorer will follow later.
In addition to those changes, project structure was improved, putting
classes that should not be used by mods directly into an Internal namespace.
This is an overview of the most important changes, please look over the code
for a full list.