Add Game class and get a basic Injector working
The Injector project is now capable of injecting the necessary
changes into Assembly-UnityScript.dll to load the mod loader.
The references in all the projects use the BESIEGE_LOCATION env
variable now.
The InternalModLoader project is removed since it's not necessary anymore.
Finally fix all target frameworks so nothing depends on .NET 4 anymore,
everything is fine with what Unity provides.