Change AssemblyName when reloading mod in debug helper, not just version.
Fixes #37.
Unity seems to use a combination of assembly name and the full name
of classes (including the namespace) when adding components.
Even if using the generic methods, it still uses the name internally.
This caused problems when mutliple versions of classes that have the
same name following this scheme were loaded by the debug helper.
Unity added the wrong version of the component.
Modifying the assembly name in addition to the version (which Unity
does not seem to consider) fixes this problem.