Fix an ArugmentException in the block loader
The first time any level is loaded, all blocks are loaded.
When the next level is loaded, they already are, and everything is fine.
On the third and later loads however, there was an ArgumentException
begin thrown in NewBlockPrefab. This fixes that, by moving the check
whether or not a block is already loaded up a bit.